chore(CreatePolicy): clean policy creation with redux forms
