--- src/SmartComponents/CreatePolicy/CreateSCAPPolicy/CreateSCAPPolicy.js
@@ -1,166 +1,0 @@
-import React from 'react';
-import propTypes from 'prop-types';
-import { compose } from 'redux';
-import { connect } from 'react-redux';
-import {
-  formValueSelector,
-  reduxForm,
-  propTypes as reduxFormPropTypes,
-} from 'redux-form';
-import {
-  Form,
-  FormGroup,
-  Spinner,
-  Content,
-  ContentVariants,
-} from '@patternfly/react-core';
-import { Tile } from '@patternfly/react-core/deprecated';
-import { TableStateProvider } from 'bastilian-tabletools';
-import useSupportedProfiles from 'Utilities/hooks/api/useSupportedProfiles';
-import useSecurityGuidesOS from 'Utilities/hooks/api/useSecurityGuidesOS';
-import {
-  StateViewPart,
-  StateViewWithError,
-  PolicyTypesTable,
-  PolicyTypeTooltip,
-} from 'PresentationalComponents';
-
-const serialiseOsVersions = (profiles = []) =>
-  profiles.map((profile) => ({
-    ...profile,
-    supportedOsVersions: profile.os_minor_versions.map(
-      (minorVersion) => `${profile.os_major_version}.${minorVersion}`,
-    ),
-  }));
-
-const CreateSCAPPolicy = ({
-  selectedOsMajorVersion,
-  selectedProfile,
-  change,
-}) => {
-  const {
-    data: { data: availableOsMajorVersions } = {},
-    error: availableOsMajorVersionsError,
-    loading: availableOsMajorVersionsLoading,
-  } = useSecurityGuidesOS();
-
-  const {
-    data: { data: availableProfiles, meta: { total } = {} } = {},
-    loading: availableProfilesLoading,
-    error: availableProfilesError,
-  } = useSupportedProfiles({
-    params: {
-      // The reason we are using the ^ operator is because Akamai is blocking
-      // this request when it's combined with title filter. See: RHINENG-23142
-      filters: `os_major_version ^ (${selectedOsMajorVersion})`,
-    },
-    useTableState: true,
-    skip: selectedOsMajorVersion === undefined,
-  });
-  const profilesData =
-    availableProfiles && serialiseOsVersions(availableProfiles);
-
-  const data =
-    availableOsMajorVersions === undefined
-      ? undefined
-      : {
-          availableOsMajorVersions,
-          availableProfiles: profilesData,
-        };
-
-  return (
-    <StateViewWithError
-      stateValues={{
-        error: availableOsMajorVersionsError || availableProfilesError,
-        data,
-        loading: availableOsMajorVersionsLoading,
-      }}
-    >
-      <StateViewPart stateKey="loading">
-        <Spinner />
-      </StateViewPart>
-      <StateViewPart stateKey="data">
-        <Content>
-          <Content component={ContentVariants.h1} className="pf-v6-u-mb-md">
-            Create SCAP policy
-          </Content>
-          <Content component="p" className="pf-v6-u-mb-md">
-            Select the operating system and policy type for this policy.
-          </Content>
-        </Content>
-        <Form>
-          <FormGroup
-            label="Operating system"
-            isRequired
-            fieldId="security-guide"
-          >
-            {(data?.availableOsMajorVersions || []).map((osMajorVersion) => (
-              <Tile
-                key={`rhel${osMajorVersion}-select`}
-                className="pf-v6-u-mr-md"
-                title={`RHEL ${osMajorVersion}`}
-                onClick={() => {
-                  change('osMajorVersion', osMajorVersion);
-                }}
-                isSelected={selectedOsMajorVersion === osMajorVersion}
-                isStacked
-              />
-            ))}
-          </FormGroup>
-          {selectedOsMajorVersion ? (
-            <FormGroup
-              isRequired
-              labelHelp={<PolicyTypeTooltip />}
-              label="Policy type"
-              fieldId="policy-type"
-            >
-              <PolicyTypesTable
-                profiles={data?.availableProfiles || []}
-                onChange={(profile) => {
-                  change('profile', profile);
-                  change('selectedRuleRefIds', undefined);
-                  change('systems', []);
-                }}
-                selectedProfile={selectedProfile}
-                loading={availableProfilesLoading}
-                total={total}
-              />
-            </FormGroup>
-          ) : (
-            <React.Fragment />
-          )}
-        </Form>
-      </StateViewPart>
-    </StateViewWithError>
-  );
-};
-
-CreateSCAPPolicy.propTypes = {
-  change: reduxFormPropTypes.change,
-  selectedProfile: propTypes.object,
-  selectedOsMajorVersion: propTypes.string,
-};
-
-const CreateSCAPPolicyTableStateProvider = (props) => {
-  return (
-    <TableStateProvider>
-      <CreateSCAPPolicy {...props} />
-    </TableStateProvider>
-  );
-};
-
-const selector = formValueSelector('policyForm');
-
-const CreateSCAPPolicyWithRedux = compose(
-  connect((state) => ({
-    selectedProfile: selector(state, 'profile'),
-    selectedOsMajorVersion: selector(state, 'osMajorVersion'),
-  })),
-  reduxForm({
-    form: 'policyForm',
-    destroyOnUnmount: false,
-    forceUnregisterOnUnmount: true,
-  }),
-)(CreateSCAPPolicyTableStateProvider);
-
-export { CreateSCAPPolicyWithRedux, CreateSCAPPolicyTableStateProvider };
