--- src/SmartComponents/CreatePolicy/CreateSCAPPolicy/index.js
@@ -1,1 +1,0 @@
-export { CreateSCAPPolicyWithRedux as default } from './CreateSCAPPolicy';
