--- src/SmartComponents/CreatePolicy/FinishedCreatePolicy/index.js
@@ -1,1 +1,0 @@
-export { default } from './FinishedCreatePolicy';
