+++ src/SmartComponents/CreatePolicy/componentMapper.js
