+++ src/SmartComponents/CreatePolicy/components/CreateSCAPPolicyStep.js
