+++ src/SmartComponents/CreatePolicy/components/FinishedStep.js
