+++ src/SmartComponents/CreatePolicy/index.js
